requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1

# Vector stores and embeddings
faiss-cpu>=1.7.4
openai>=1.0.0

# Document loaders
pypdf>=3.17.0
unstructured>=0.11.0
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Streamlit
streamlit>=1.28.0

# Environment variables
python-dotenv>=1.0.0

# Additional utilities
tiktoken>=0.5.1